v2 (rounds 25-34): seats, counted barrier, four ships, lobby, host migration,
host-ruled deaths, de-jitter, netrig -- device-confirmed at two players, LAN
and online.

v3 (rounds 35-38): -Werror clean, ARC, launch screen, the Metal backend as the
one renderer (OpenGL ES retired), ink menu buttons, centered text, the
tutorial's BACK off the title card. Device-confirmed on build 225: "tout est ok".

+++ engine/src/sounds.c
@@ -48,7 +48,7 @@ void SND_Load(char* filename,int soundID)
 	Log_Printf("[SND_Load] Loaded sound: %s sample_rate=%ld ",filename,sound->metaData.sample_rate);
 	
 	// Calculate buffer size
-	sound->size = sound->metaData.samples * sound->metaData.sample_size * sound->metaData.channels;
+	sound->size = (int)(sound->metaData.samples * sound->metaData.sample_size * sound->metaData.channels);
 	
 	if( sound->metaData.sample_size == 2 )
 	{
